requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
django-geojson==3.2.0
django-leaflet==0.28.2
django-location-field==2.1.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-gis==0.17
gunicorn==20.1.0
idna==3.3
jsonfield==3.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
